Add -Wconversion casts to pam-util
Add casts to the pam-util library in places where signed to unsigned
conversions have been confirmed to not be bugs, in preparation for
using -Wconversion.  This warning flag is still unused by default due
a bug in handling htons.

+++ pam-util/options.c
@@ -410,7 +410,7 @@ option_compare(const void *key, const void *member)
     if (p == NULL)
         return strcmp(string, option->name);
     else {
-        length = (p - string);
+        length = (size_t) (p - string);
         if (length == 0)
             return -1;
         result = strncmp(string, option->name, length);
